pyproject.toml
[project]
name = "ai-buddy"
version = "0.1.0"
description = "A tool using AI script"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lmstudio>=1.2.0",
    "notion-client>=2.3.0",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
    "transformers>=4.50.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
